+++ ct_cannabis_names.py
@@ -27,7 +27,6 @@ import csv
 import os
 import re
 import sys
-import json
 
 try:
     import requests